+++ src/main/java/com/zerobase/moy/controller/DiaryController.java
@@ -3,8 +3,8 @@ package com.zerobase.moy.controller;
 import com.zerobase.moy.data.entity.User;
 import com.zerobase.moy.data.model.diary.DiaryForm;
 import com.zerobase.moy.data.model.diary.DiaryResultDto;
-import com.zerobase.moy.data.model.response.ApiResponse;
-import com.zerobase.moy.data.model.response.ResponseCode;
+import com.zerobase.moy.response.ApiResponse;
+import com.zerobase.moy.response.ResponseCode;
 import com.zerobase.moy.service.DiaryService;
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
