+++ src/main/java/com/zerobase/moy/controller/SignController.java
@@ -1,7 +1,7 @@
 package com.zerobase.moy.controller;
 
-import com.zerobase.moy.data.model.response.ApiResponse;
-import com.zerobase.moy.data.model.response.ResponseCode;
+import com.zerobase.moy.response.ApiResponse;
+import com.zerobase.moy.response.ResponseCode;
 import com.zerobase.moy.data.model.sign.SignInForm;
 import com.zerobase.moy.data.model.sign.SignUpForm;
 import com.zerobase.moy.service.SignService;
