+++ src/main/java/com/zerobase/moy/service/impl/DiaryServiceImpl.java
@@ -4,8 +4,8 @@ import com.zerobase.moy.data.entity.Diary;
 import com.zerobase.moy.data.entity.User;
 import com.zerobase.moy.data.model.diary.DiaryForm;
 import com.zerobase.moy.data.model.diary.DiaryRequestDto;
-import com.zerobase.moy.exception.CustomException;
-import com.zerobase.moy.exception.ErrorCode;
+import com.zerobase.moy.response.exception.CustomException;
+import com.zerobase.moy.response.exception.ErrorCode;
 import com.zerobase.moy.repository.DiaryRepository;
 import com.zerobase.moy.service.DiaryService;
 import java.util.Objects;
